Add the is_admin decorator as well as the is_pkg_admin and is_pkgdb_admin methods
The @is_admin decorator is meant to restrict access to the page only accessible to
pkgdb admin, for example: create a collection or a package.

The is_pkgdb_admin is a simple function returning a boolean whether the user is part
of the admin FAS group set in the configuration file.

The is_pkg_admin is returns boolean whether the user has approveacls rights on a
specific branch of a specific package.