Update packages namespace
When displaying the list of packages, the option to create a new packages is only
visible to the pkgdb admins.

Restrict the new packages page to pkgdb admins.

Handle case the package does not exists, to display an error

Retrieve for each branch of a package if the current user has 'approveacls' rights, so that
we can display the link to update ACLs only if they are allowed.