More personalisation of JSON returned
Using package=True/False one can specify if the PackageListing json representation
should also include a representation of the package object it's linked to.
This allows reducing the amount of duplicated information in the JSON output of some API calls.

- Using collection=['<branchname>'] one may restrict the representation of
the ACLs in JSON to only the specified branches, thus omitting the other branches.