Include in the /api/vcs output packages that are orphaned but have no maintainer left
Basically, the first query retrieves all the packages in active branches
that are Active or Orphaned and the person having Approved commit ACL
on said branches.

The problem is that some package are orphaned and all the maintainers
dropped their commit ACLs. This means that the package will not show up
which means, no-one can commit to it.

That is not something we want since we actually want to allow provenpackagers
to commit on orphaned packages.

This change allows this but retrieving the list of active packages on
active branches and adding the ones missing from the results of the first
query to the data returned.

Fixes https://github.com/fedora-infra/pkgdb2/issues/86