Adjust the update_pkg_poc method to allow restricting its action to a specified user
This way we can ensure we only orphan the package on branches where the
user is POC even if it's not the case for the specified branch.

Useful for mass-orphan where we orphan all branches at once and thus
the list of branches includes more than just what the user has.