Retrieve the content of the dead.package file when clicking on 'Retired'
On the top right of the page is the table listing the different status
of the package for each collection.
If that status says 'Retired' we underline it and make it so that when
clicking on it we `alert()` the user with the content of that file.

Fixes https://github.com/fedora-infra/pkgdb2/issues/119