Add support for user notification
This change retrieves from the error message the email of the user
affected.
It then sends an email to that user about the situation and the steps
that can be taken to resolve it.
The email is then stored in a JSON object on disc, together with the
current date (UTC) as unix timestamp.
For every run, the script will check the date of the last update (ie last
time we notified the user about the issue) and if the delta is more than
3600 seconds (60*60 = 1 hour), a new email will be sent.

So users will receive a hourly reminder to change their synchronize their
bugzilla/FAS emails/account while admins will keep getting the information
every 10 minutes (as currently).