Adjust the bug API endpoint for package namespace
We currently had two collections: Fedora and Fedora EPEL.
With the landing of the namespace feature, we now have more collections
to sync to bugzilla. So for each namespace that is not the default
`rpms`, we append the namespace (capitalized) to the collection name.
As such, docker namespace on Fedora becomes 'Fedora Docker'.