Create a fedmsg consumer to update bugzilla when PoC changes
When changing the point of contact (PoC) for a component having a lots
of bugs opened (think dnf or kernel), we need to update each and every
bugs to be assigned to the new PoC.

This can take a while and actually lead to a time-out request on pkgdb.

In order to prevent this, we're making the process to update bugzilla
async of the request changing the PoC and relying on fedmsg. This will
allow us to update the PoC on the one side and update bugzilla on the
other side without the risk of running into time-outs.