Added Support for dev environment under Vagrant
This adds a Vagrant file that allows a developer to quickly
set up a dev environment of pkgdb2 using the infra db dump.
It uses vagrant-libvirt and vagrantsshfs for file syncing.

Also in this change, added a new param in the runserver.py
that allows the host to be set as a commandline param.