update vagrant ansible role to upgrade database
Previously, the vagrant config did not upgrade the database
when provisioning the new machine. This could lead to issues
if a database change was commited to master, and we pull the
dbdump from prod.

This change copies the alembic.ini from the utility directory,
changes the dburl to what the vagrant setup expects, and runs
the alembic upgrade command.